pyproject.toml
[project]
name = "ai-textbook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"mkdocs==1.6.1",
	"mkdocs-material==9.6.16",
	"pymdown-extensions==10.16.1",
	"mkdocs-kroki-plugin==0.9.0",
	"mkdocs-include-markdown-plugin==7.1.6",
	"mkdocs-macros-plugin==1.3.9",
	"mkdocs-git-revision-date-localized-plugin==1.4.7",
	"mkdocs-awesome-nav==3.1.2",
]

[project.optional-dependencies]
# Development extras: use `pip install -e .[dev]` to install the following tools
dev = [
	"annotated-types==0.7.0",
	"babel==2.17.0",
	"backrefs==5.9",
	"beautifulsoup4==4.13.4",
	"bracex==2.6",
	"certifi==2025.7.14",
	"charset-normalizer==3.4.2",
	"click==8.2.1",
	"colorama==0.4.6",
	"defusedxml==0.7.1",
	"editorconfig==0.17.1",
	"gitdb==4.0.12",
	"gitpython==3.1.45",
	"hjson==3.1.0",
	"idna==3.10",
	"jinja2==3.1.6",
	"jsbeautifier==1.15.4",
	"markdown==3.8.2",
	"markupsafe==3.0.2",
	"mergedeep==1.3.4",
	"mkdocs==1.6.1",
	"mkdocs-awesome-nav==3.1.2",
	"mkdocs-get-deps==0.2.0",
	"mkdocs-git-revision-date-localized-plugin==1.4.7",
	"mkdocs-include-markdown-plugin==7.1.6",
	"mkdocs-kroki-plugin==0.9.0",
	"mkdocs-macros-plugin==1.3.9",
	"mkdocs-material==9.6.16",
	"mkdocs-material-extensions==1.3.1",
	"mkdocs-mermaid2-plugin==1.2.1",
	"mkdocs-plantuml==0.1.1",
	"natsort==8.4.0",
	"packaging==25.0",
	"paginate==0.5.7",
	"pathspec==0.12.1",
	"pip==25.2",
	"platformdirs==4.3.8",
	"pydantic==2.11.7",
	"pydantic-core==2.33.2",
	"pygments==2.19.2",
	"pymarkdown==0.1.4",
	"pymdown-extensions==10.16.1",
	"python-dateutil==2.9.0.post0",
	"pytz==2025.2",
	"pyyaml==6.0.2",
	"pyyaml-env-tag==1.1",
	"requests==2.32.4",
	"result==0.17.0",
	"setuptools==80.9.0",
	"six==1.17.0",
	"smmap==5.0.2",
	"soupsieve==2.7",
	"super-collections==0.5.3",
	"termcolor==3.1.0",
	"toolz==1.0.0",
	"typing-extensions==4.14.1",
	"typing-inspection==0.4.1",
	"urllib3==2.5.0",
	"watchdog==6.0.0",
	"wcmatch==10.1",
	"wheel==0.45.1",
]
